Convert launchdev.js script to playwright loader
